requirements.txt
python-telegram-bot==20.0b1
openai
